Move SegmenterGui utils into a separate file
